requirements/prod.txt
-r base.txt

gunicorn==0.17.2

django-heroku-redisify==0.2.1
django-redis-cache==0.9.7.obc1
hiredis==0.1.1
redis==2.7.5
